+++ packages/yourloops/components/patient-list/models/enums/patient-list.enum.ts
@@ -49,6 +49,7 @@ export enum PatientListColumns {
 
 export enum PendingPatientListColumns {
   Actions = 'actions',
-  Icon = 'icon',
-  Patient = 'patient'
+  Date = 'date',
+  Email = 'email',
+  InviteSentBy = 'invite-sent-by',
 }