+++ packages/yourloops/components/patient-list/models/grid-row.model.ts
@@ -25,7 +25,7 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-import { type PatientListColumns } from './enums/patient-list.enum'
+import { type PatientListColumns, type PendingPatientListColumns } from './enums/patient-list.enum'
 import { type Patient } from '../../../lib/patient/models/patient.model'
 
 export interface GridRowModel {
@@ -44,4 +44,13 @@ export interface GridRowModel {
   [PatientListColumns.Variance]?: number
   [PatientListColumns.LastDataUpdate]?: string
   [PatientListColumns.Actions]: Patient
+}
+
+export interface PendingGridRowModel {
+  id: string
+  isInviteCreatedByLoggedUser: boolean
+  [PendingPatientListColumns.Actions]: Patient
+  [PendingPatientListColumns.Date]: string
+  [PendingPatientListColumns.Email]: string
+  [PendingPatientListColumns.InviteSentBy]: string
 }