+++ packages/yourloops/layout/caregiver-layout.tsx
@@ -34,7 +34,7 @@ import InvalidRoute from '../components/invalid-route'
 import ProfilePage from '../pages/profile/profile-page'
 import NotificationsPage from '../pages/notifications'
 import { AppUserRoute } from '../models/enums/routes.enum'
-import { PatientList } from '../components/patient-list/patient-list'
+import { PatientListPage } from '../components/patient-list/patient-list-page'
 import { PatientListProvider } from '../lib/providers/patient-list.provider'
 
 export const CaregiverLayout: FunctionComponent = () => {
@@ -47,7 +47,7 @@ export const CaregiverLayout: FunctionComponent = () => {
             <Route path={AppUserRoute.Preferences} element={<ProfilePage />} />
             <Route path={AppUserRoute.Notifications} element={<NotificationsPage />} />
             <Route path={`${AppUserRoute.Patient}/:patientId/*`} element={<PatientDataPage />} />
-            <Route path={AppUserRoute.Home} element={<PatientList />} />
+            <Route path={AppUserRoute.Home} element={<PatientListPage />} />
             <Route
               path="/"
               element={<Navigate to={AppUserRoute.Home} replace />}